# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Heymaty - OpenAPI 3.0",
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://heymaty.com/"
    }
  ],
...
  "paths": {
...
    "/comments/post/{id}": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "get a comment",
        "description": "get a comment",
        "operationId": "getComment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "object",
                      "properties": {
                        "docs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Comments"
                          }
                        },
                        "totalDocs": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "pagingCounter": {
                          "type": "number"
                        },
                        "hasPrevPage": {
                          "type": "boolean"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "prevPage": {
                          "type": "number"
                        },
                        "nextPage": {
                          "type": "number"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/comments/count/post/{id}": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "get a count comments",
        "description": "get a count comments",
        "operationId": "getCountComments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
...
    "/posts/all-public-posts": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Retrieve all public posts",
        "description": "Get all public posts based on the given filters and pagination.",
        "operationId": "getFeedPosts",
        "parameters": [
          {
            "in": "query",
            "name": "data",
            "required": true,
            "description": "Filtering data for retrieving public posts.",
            "schema": {
              "type": "string",
              "example": "{\"name\": \"D182\", \"ages\": [18,30], \"languages\": [], \"cities\": [\"654115abf51d393d9341c0e9\"], \"country\": \"654115abf51d393d9341c0e6\"}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/posts"
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "length": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
    "/posts/user/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get all girl post route",
        "description": "Get all girl post route",
        "operationId": "getGirlPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Write user _id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/posts"
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Посты получены"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      }
    },
    "/posts/": {
      "post": {
        "tags": [
          "posts"
        ],
        "summary": "Create new post route",
        "description": "Create new post route",
        "operationId": "createPost",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  },
                  "videos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of video files"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "$ref": "#/components/schemas/posts"
                    },
                    "message": {
                      "type": "string",
                      "example": "Пост успешно создан"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get single post route",
        "description": "Get single post route",
        "operationId": "getPost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Write post _id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "$ref": "#/components/schemas/posts"
                    },
                    "message": {
                      "type": "string",
                      "example": "Пост получен"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "post not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl",
          "user"
        ]
      },
      "patch": {
        "tags": [
          "posts"
        ],
        "summary": "Update item route",
        "description": "Update item route",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Write post _id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "existedImages": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  },
                  "existedVideos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of image files"
                  },
                  "videos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Array of video files"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "isPublic": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "posts": {
                      "$ref": "#/components/schemas/posts"
                    },
                    "message": {
                      "type": "string",
                      "example": "Пост успешно создан"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      },
      "delete": {
        "tags": [
          "posts"
        ],
        "summary": "Delete item route",
        "description": "Delete item route",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Write post _id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Пост удален"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Пост не существует",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Вы не авторизованы",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ],
        "x-roles": [
          "girl"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "msg": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "errors"
        ]
      },
      "ErrorsResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "description": "Error message."
                },
                "param": {
                  "type": "string",
                  "description": "The field that triggered the error."
                },
                "location": {
                  "type": "string",
                  "description": "Location where the error occurred."
                }
              },
              "required": [
                "msg",
                "param",
                "location"
              ]
            }
          }
        },
        "required": [
          "errors"
        ]
      },
...
      "Comments": {
        "required": [
          "postId",
          "userId",
          "comment"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "userId": {
            "$ref": "#/components/schemas/User"
          },
          "postId": {
            "$ref": "#/components/schemas/posts"
          },
          "comment": {
            "type": "string",
            "maxLength": 200
          },
          "mentions": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
      "posts": {
        "required": [
          "title",
          "description"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "userId": {
            "$ref": "#/components/schemas/User"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "videos": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}